requirements.txt
transformers>=4.40.0
torch>=2.0.0
hf_xet>=0.4.1
openai>=1.96.0
python-dotenv>=1.0.0
ollama>=0.0.1